test: update executor test
