+++ tests/jax/integration/executor_test.py
@@ -17,105 +17,49 @@
 
 import functools
 import time
-from datetime import datetime
-from typing import Any
 
-import launchpad as lp
-import optax
 import pytest
 from launchpad.launch.test_multi_threading import (
     address_builder as test_address_builder,
 )
 
-from mava.components.jax.building.data_server import OnPolicyDataServer
-from mava.components.jax.building.distributor import Distributor
-from mava.components.jax.building.parameter_client import ExecutorParameterClient
-from mava.components.jax.updating.parameter_server import DefaultParameterServer
-from mava.specs import DesignSpec
-from mava.systems.jax import ippo
-from mava.systems.jax.ippo.components import ExtrasLogProbSpec
-from mava.systems.jax.system import System
-
+from mava.systems.jax import System
+from mava.types import OLT
 from mava.utils.environments import debugging_utils
-from mava.utils.loggers import logger_utils
+from tests.jax.integration.mock_systems import (
+    mock_system_multi_thread,
+    mock_system_single_process,
+)
+
+# Environment.
+environment_factory = functools.partial(
+    debugging_utils.make_environment,
+    env_name="simple_spread",
+    action_space="discrete",
+)
 
 
 @pytest.fixture
-def test_system() -> System
-    """Mappo system"""
-    return mappo.MAPPOSystem()
+def test_system_sp() -> System:
+    """A single process built system"""
+    return mock_system_single_process()
 
 
-def test_executor_single_process_with_adder(test_system: System) -> None:
+@pytest.fixture
+def test_system_mt() -> System:
+    """A multi thread built system"""
+    return mock_system_multi_thread()
+
+
+def test_executor_single_process_with_adder(test_system_sp: System) -> None:
     """Test if the executor instantiates processes as expected."""
-
-    # Environment.
-    environment_factory = functools.partial(
-        debugging_utils.make_environment,
-        env_name="simple_spread",
-        action_space="discrete",
-    )
-
-    # Networks.
-    def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(32, 32),
-            critic_layer_sizes=(64, 64),
-            *args,
-            **kwargs,
-        )
-
-
-    # Checkpointer appends "Checkpoints" to checkpoint_dir.
-    base_dir = "~/mava"
-    mava_id = str(datetime.now())
-    checkpoint_subpath = f"{base_dir}/{mava_id}"
-    # Log every [log_every] seconds.
-    log_every = 1
-    logger_factory = functools.partial(
-        logger_utils.make_logger,
-        directory=base_dir,
-        to_terminal=True,
-        to_tensorboard=True,
-        time_stamp=mava_id,
-        time_delta=log_every,
-    )
-    # Optimizer.
-    optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0),
-        optax.adam(1e-4),
-    )
-
-    # Build the test_system
-    test_system.build(
-        environment_factory=environment_factory,
-        network_factory=network_factory,
-        logger_factory=logger_factory,
-        experiment_path=checkpoint_subpath,
-        optimizer=optimizer,
-        executor_parameter_update_period=10,
-        multi_process=False,
-        run_evaluator=True,
-        num_executors=1,
-        use_next_extras=False,
-        max_queue_size=5000,
-        sample_batch_size=2,
-        max_in_flight_samples_per_worker=4,
-        num_workers_per_iterator=-1,
-        rate_limiter_timeout_ms=-1,
-        nodes_on_gpu=[],
-        lp_launch_type=lp.LaunchType.TEST_MULTI_THREADING,
-        sequence_length=4,
-        period=4,
-    )
-
     (
         data_server,
         parameter_server,
         executor,
         evaluator,
         trainer,
-    ) = test_system._builder.store.system_build
+    ) = test_system_sp._builder.store.system_build
 
     # _writer.append needs to be called once to get _writer.history
     # _writer.append called in observe_first and observe
@@ -167,75 +111,16 @@ def test_executor_single_process_with_adder(test_system: System) -> None:
     )
 
 
-def test_executor_single_process_without_adder(test_system: System) -> None:
+def test_executor_single_process_without_adder(test_system_sp: System) -> None:
     """Test if the executor instantiates processes as expected."""
-    # Environment.
-    environment_factory = functools.partial(
-        debugging_utils.make_environment,
-        env_name="simple_spread",
-        action_space="discrete",
-    )
-
-    # Networks.
-    def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(32, 32),
-            critic_layer_sizes=(64, 64),
-            *args,
-            **kwargs,
-        )
-
-
-    # Checkpointer appends "Checkpoints" to checkpoint_dir.
-    base_dir = "~/mava"
-    mava_id = str(datetime.now())
-    checkpoint_subpath = f"{base_dir}/{mava_id}"
-    # Log every [log_every] seconds.
-    log_every = 1
-    logger_factory = functools.partial(
-        logger_utils.make_logger,
-        directory=base_dir,
-        to_terminal=True,
-        to_tensorboard=True,
-        time_stamp=mava_id,
-        time_delta=log_every,
-    )
-    # Optimizer.
-    optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0),
-        optax.adam(1e-4),
-    )
-
-    # Build the test_system
-    test_system.build(
-        environment_factory=environment_factory,
-        network_factory=network_factory,
-        logger_factory=logger_factory,
-        experiment_path=checkpoint_subpath,
-        optimizer=optimizer,
-        executor_parameter_update_period=10,
-        multi_process=False,
-        run_evaluator=True,
-        num_executors=1,
-        use_next_extras=False,
-        max_queue_size=5000,
-        sample_batch_size=2,
-        max_in_flight_samples_per_worker=4,
-        num_workers_per_iterator=-1,
-        rate_limiter_timeout_ms=-1,
-        nodes_on_gpu=[],
-        lp_launch_type=lp.LaunchType.TEST_MULTI_THREADING,
-        sequence_length=4,
-        period=4,
-    )
-
     (
         data_server,
         parameter_server,
         executor,
         evaluator,
         trainer,
-    ) = test_system._builder.store.system_build
+    ) = test_system_sp._builder.store.system_build
+
     # Remove adder
     executor._executor.store.adder = None
     # Run an episode
@@ -270,67 +155,14 @@ def test_executor_single_process_without_adder(test_system: System) -> None:
     assert not hasattr(executor._executor.store.adder, "add")
 
 
-def test_executor_multi_process_with_adder(test_system: System) -> None:
+def test_executor_multi_thread_with_adder(test_system_mt: System) -> None:
     """Test if the executor instantiates processes as expected."""
-    # Environment.
-    environment_factory = functools.partial(
-        debugging_utils.make_environment,
-        env_name="simple_spread",
-        action_space="discrete",
-    )
-
-    # Networks.
-    def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(32, 32),
-            critic_layer_sizes=(64, 64),
-            *args,
-            **kwargs,
-        )
-
-    # Checkpointer appends "Checkpoints" to checkpoint_dir.
-    base_dir = "~/mava"
-    mava_id = str(datetime.now())
-    checkpoint_subpath = f"{base_dir}/{mava_id}"
-
-    # Log every [log_every] seconds.
-    log_every = 10
-    logger_factory = functools.partial(
-        logger_utils.make_logger,
-        directory=base_dir,
-        to_terminal=True,
-        to_tensorboard=True,
-        time_stamp=mava_id,
-        time_delta=log_every,
-    )
-
-    # Optimizer.
-    optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0),
-        optax.adam(1e-4),
-    )
-
-
-    # Build the system
-    test_system.build(
-        environment_factory=environment_factory,
-        network_factory=network_factory,
-        logger_factory=logger_factory,
-        experiment_path=checkpoint_subpath,
-        optimizer=optimizer,
-        executor_parameter_update_period=1,
-        multi_process=True,
-        run_evaluator=True,
-        num_executors=1,
-        max_queue_size=500,
-        use_next_extras=False,
-        sample_batch_size=5,
-        nodes_on_gpu=[],
-        lp_launch_type=lp.LaunchType.TEST_MULTI_THREADING,
-    )
-    (executor_node,) = test_system._builder.store.program._program._groups["executor"]
+    (executor_node,) = test_system_mt._builder.store.program._program._groups[
+        "executor"
+    ]
     test_address_builder.bind_addresses([executor_node])
-    test_system.launch()
+
+    test_system_mt.launch()
     time.sleep(10)
     executor = executor_node._construct_instance()
